Updated some.
2003-05-19  Jeffrey Stedfast  <fejj@ximian.com>

	* docs/reference/*: Updated some.

	* gmime/gmime-multipart-signed.c: Same as below.

	* gmime/gmime-multipart-encrypted.c: Fixed the prototypes for
	write_to_stream().

	* gmime/gmime-gpg-context.c (gpg_ctx_new): Initialise seen_eof1 to
	TRUE.
	(gpg_ctx_set_ostream): Set seen_eof1 to FALSE here, this way we
	only ever check for EOF for gpg's stdout if we care about the data
	received from it.
	(gpg_ctx_parse_status): Remove the hack that set seen_eof1 to TRUE
	when we received a TRUST status from gpg when verifying a
	signature. Same for when importing signatures.
	(gpg_ctx_op_step): Only FD_SET() fd's that we have not finished
	reading from.
	(gpg_ctx_op_exited): #if 0'd for now since we aren't using it
	anymore. Keeping it around for historical reasons and in case we
	find we really do need it back.

+++ gmime/gmime-multipart-encrypted.c
@@ -54,7 +54,7 @@ static const char *multipart_encrypted_get_header (GMimeObject *object, const ch
 static void multipart_encrypted_remove_header (GMimeObject *object, const char *header);
 static void multipart_encrypted_set_content_type (GMimeObject *object, GMimeContentType *content_type);
 static char *multipart_encrypted_get_headers (GMimeObject *object);
-static int multipart_encrypted_write_to_stream (GMimeObject *object, GMimeStream *stream);
+static ssize_t multipart_encrypted_write_to_stream (GMimeObject *object, GMimeStream *stream);
 
 
 static GMimeMultipartClass *parent_class = NULL;
